agents/problem_solver_agent/requirements.txt
openai==1.76.0
python-dotenv==1.1.0
uagents==0.22.3
uagents_core==0.2.3
